created entities for projects/blog dashboards to clean up the admin dashboard file
